// src/app/api/model/sectionsTurnoutDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SectionTurnoutDto } from './sectionTurnoutDto';


export interface SectionsTurnoutDto { 
    dateElectors?: string;
    totalElectors?: number;
    votingElectors?: number;
    sectionTurnoutDtoList?: Array<SectionTurnoutDto>;
}